+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   scroll-capable components and can be overridden per-component
 - Message popup now supports scrolling with a scrollbar
 - Command popup output now preserves ANSI color
+- Drag to resize pane divider in all tabs
 
 ## [0.8.0] - 2026-04-19